+++ apps/main/lib/blog/main/view/context.rb
@@ -1,0 +1,10 @@
+require "blog/view/context"
+
+module Blog
+  module Main
+    module View
+      class Context < Blog::View::Context
+      end
+    end
+  end
+end
