+++ apps/main/lib/blog/main/view/controller.rb
@@ -1,0 +1,19 @@
+# auto_register: false
+
+require "slim"
+require "dry/view/controller"
+require "blog/main/container"
+
+module Blog
+  module Main
+    module View
+      class Controller < Dry::View::Controller
+        configure do |config|
+          config.paths = [Container.root.join("web/templates")]
+          config.context = Container["view.context"]
+          config.layout = "application"
+        end
+      end
+    end
+  end
+end
