+++ apps/main/lib/blog/main/view/parts/article.rb
@@ -1,0 +1,16 @@
+require "dry/view/part"
+require "commonmarker"
+
+module Blog
+  module Main
+    module View
+      module Parts
+        class Article < Dry::View::Part
+          def body_html
+            @body_html ||= CommonMarker.render_doc(body, :SMART).to_html
+          end
+        end
+      end
+    end
+  end
+end
