--- apps/main/lib/main/view/context.rb
@@ -1,8 +1,0 @@
-require "blog/view/context"
-
-module Main
-  module View
-    class Context < Blog::View::Context
-    end
-  end
-end
