+++ apps/main/system/blog/main/import.rb
@@ -1,0 +1,7 @@
+require_relative "container"
+
+module Blog
+  module Main
+    Import = Main::Container.injector
+  end
+end
