--- apps/main/system/main/import.rb
@@ -1,5 +1,0 @@
-require_relative "container"
-
-module Main
-  Import = Main::Container.injector
-end
